Remove IS calculation, move it to separate tool
